In this commit I made hnparser into a blocking server. The code waits for incoming connections and processes what these connections give it. Why? I feel that the scraping of content should take place outside of the hacker-archives site and this was one method of accomplishing that.
Another change made to the code was adding print statements which explain what part of link processing
is taking place. I originally added these to help with debugging, but I'm leaving them since I don't
see a good reason not to.

The last notable chance made in this commit is syntactical in nature. I found out on IRC that let and
defn surround code in an implicit do. I also found out that when is more idiomatic than if given that
the if expression has no else clause. My code was updated accordingly.