Makes lein install work.
We were well behind the latest Clojure release. This
brings us forward in time where we can take advantage
of modern Clojure libraries.

This project was originally intended to handle the use
case of scraping hacker news. Since then I've realized
that scraping capabilities are needed in a more general
way. So I'm not building a Hacker News scraper, but a
scraper. I'm also not building a web scraper, but a scraper.
As in, I'm building something which can walk graphs in a
general sense rather than walking web graphs in a specific
sense.

The motivation behind this is to provide capabilities to
myself so my content isn't trapped within the tools I happen
to use.

I'm planning to eventually make this tool I build for myself
available to others as well. It is a fairly simple tool, but
it will likely be valuable to others.