Step toward 12 factor application.
Right now this project is a cross between an old fanfiction.net
story parsing code base and a hacker news scraper. What I'm
working toward making it is a tool for doing graph processing on
arbitrary graphs.

When I say arbitrary graphs what I mean to imply is that this is
not intended to be a scraper that can only scrape web pages though
it will of course be able to do that. Instead it is a scraper that
can scrape many different types of graph structures. This could
mean scraping the exports of a Notion database of the save file of
an Athens Research database. It could mean scraping the web. It
could even mean doing a walk over a game tree.

Right now that isn't what this is and a first step to getting it
to the point where it is this is getting things set up so that the
scraping can be carried out in a scalable manner. In the short term
that means getting the existing code for doing walks and scrapes of
particular web pages back up and running and saving things.

One of the most complex parts of projects is starting them up.
For humans that takes about nine months and then has considerable
time spent during learning. The 12 factor app approach dramatically
lowers the cognitive burden to getting an application stood up.
So I'm setting up the application to be containerized in such a way
as to enable a 12 factor application compatible standup pattern.